requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
httpx==0.26.0
mercadopago==2.2.1
pydantic==2.5.3
python-multipart==0.0.6
